chore: upgrade target sdk version and gradle tools.
