+++ SignaturePad-Example/src/main/res/layout/activity_main.xml
@@ -33,7 +33,7 @@
                 android:layout_alignParentBottom="true"
                 android:text="@string/agreement" />
 
-        <com.prologapp.signaturepad.views.SignaturePad
+        <com.miskonikitast.signaturepad.views.SignaturePad
             android:id="@+id/signature_pad"
             android:layout_width="fill_parent"
             android:layout_height="fill_parent" />
