+++ gradle.properties
@@ -34,8 +34,8 @@ POM_DEVELOPER_ID=gcacace
 POM_DEVELOPER_NAME=Gianluca Cacace
 POM_DEVELOPER_EMAIL=gianluca.cacace@gmail.com
 
-ANDROID_BUILD_TARGET_SDK_VERSION=29
-ANDROID_BUILD_SDK_VERSION=29
+ANDROID_BUILD_TARGET_SDK_VERSION=33
+ANDROID_BUILD_SDK_VERSION=33
 
 android.useAndroidX=true
 android.enableJetifier=true
