+++ signature-pad/src/main/AndroidManifest.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.prologapp.signaturepad"
-          android:versionCode="1"
+    android:versionCode="1"
           android:versionName="1.0.0">
 
     <application />
