+++ signature-pad/src/main/java/com/miskonikitast/signaturepad/utils/Bezier.java
@@ -1,4 +1,4 @@
-package com.prologapp.signaturepad.utils;
+package com.miskonikitast.signaturepad.utils;
 
 public class Bezier {
